+++ api/src/main/java/io/strimzi/api/kafka/model/rebalance/KafkaRebalance.java
@@ -39,7 +39,8 @@ import java.util.function.Predicate;
         group = KafkaRebalance.RESOURCE_GROUP,
         scope = KafkaRebalance.SCOPE,
         versions = {
-            @Crd.Spec.Version(name = Constants.V1BETA2, served = true, storage = true)
+            @Crd.Spec.Version(name = Constants.V1, served = true, storage = false),
+            @Crd.Spec.Version(name = Constants.V1BETA2, served = true, storage = true, deprecated = true, deprecationWarning = "Version v1beta2 of the KafkaRebalance API is deprecated. Please use the v1 version instead.")
         },
         subresources = @Crd.Spec.Subresources(
             status = @Crd.Spec.Subresources.Status()
@@ -78,7 +79,7 @@ public class KafkaRebalance extends CustomResource<KafkaRebalanceSpec, KafkaReba
     private static final long serialVersionUID = 1L;
 
     public static final String SCOPE = Constants.SCOPE_NAMESPACED;
-    public static final List<String> VERSIONS = List.of(Constants.V1BETA2);
+    public static final List<String> VERSIONS = List.of(Constants.V1, Constants.V1BETA2);
     public static final String RESOURCE_KIND = "KafkaRebalance";
     public static final String RESOURCE_LIST_KIND = RESOURCE_KIND + "List";
     public static final String RESOURCE_GROUP = Constants.RESOURCE_GROUP_NAME;