+++ api/src/main/java/io/strimzi/api/kafka/model/user/KafkaUser.java
@@ -39,9 +39,10 @@ import java.util.function.Predicate;
         group = KafkaUser.RESOURCE_GROUP,
         scope = KafkaUser.SCOPE,
         versions = {
-            @Crd.Spec.Version(name = Constants.V1BETA2, served = true, storage = true),
-            @Crd.Spec.Version(name = Constants.V1BETA1, served = true, storage = false),
-            @Crd.Spec.Version(name = Constants.V1ALPHA1, served = true, storage = false)
+            @Crd.Spec.Version(name = Constants.V1, served = true, storage = false),
+            @Crd.Spec.Version(name = Constants.V1BETA2, served = true, storage = true, deprecated = true, deprecationWarning = "Version v1beta2 of the KafkaUser API is deprecated. Please use the v1 version instead."),
+            @Crd.Spec.Version(name = Constants.V1BETA1, served = true, storage = false, deprecated = true, deprecationWarning = "Version v1beta1 of the KafkaUser API is deprecated. Please use the v1 version instead."),
+            @Crd.Spec.Version(name = Constants.V1ALPHA1, served = true, storage = false, deprecated = true, deprecationWarning = "Version v1alpha1 of the KafkaUser API is deprecated. Please use the v1 version instead.")
         },
         subresources = @Crd.Spec.Subresources(
             status = @Crd.Spec.Subresources.Status()
@@ -79,13 +80,13 @@ import java.util.function.Predicate;
 @JsonPropertyOrder({"apiVersion", "kind", "metadata", "spec", "status"})
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
-@Version(Constants.V1BETA2)
+@Version(Constants.V1)
 @Group(Constants.RESOURCE_GROUP_NAME)
 public class KafkaUser extends CustomResource<KafkaUserSpec, KafkaUserStatus> implements Namespaced, UnknownPropertyPreserving {
     private static final long serialVersionUID = 1L;
 
     public static final String SCOPE = Constants.SCOPE_NAMESPACED;
-    public static final List<String> VERSIONS = List.of(Constants.V1BETA2, Constants.V1BETA1, Constants.V1ALPHA1);
+    public static final List<String> VERSIONS = List.of(Constants.V1, Constants.V1BETA2, Constants.V1BETA1, Constants.V1ALPHA1);
     public static final String RESOURCE_KIND = "KafkaUser";
     public static final String RESOURCE_LIST_KIND = RESOURCE_KIND + "List";
     public static final String RESOURCE_GROUP = Constants.RESOURCE_GROUP_NAME;